+++ src/types/files.ts
@@ -5,6 +5,7 @@ export type DownloadBundle = {
 	project_id: string;
 	bundle_url: string;
 	branch?: string;
+	responseTooBig?: boolean;
 };
 
 export type FileDeleted = {